refactor: #61 - AuditingAt 제거 및 createdAt 필드 추가
AuditingAt에는 생성 일시와 수정 일시가 포함되어 있지만, 현재 chat_message, chat_room 테이블에서는 생성 일시만 사용됨
