+++ src/main/java/freshtrash/freshtrashbackend/repository/ChatMessageRepository.java
@@ -5,6 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ChatMessageRepository extends JpaRepository<ChatMessage, Long> {
-}
-
+public interface ChatMessageRepository extends JpaRepository<ChatMessage, Long> {}
