+++ src/main/java/freshtrash/freshtrashbackend/repository/ChatRoomRepository.java
@@ -1,4 +1,8 @@
-package freshtrash.freshtrashbackend.repository.projections;
+package freshtrash.freshtrashbackend.repository;
 
-public interface ChatRoomRepository {
-}
+import freshtrash.freshtrashbackend.entity.ChatRoom;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ChatRoomRepository extends JpaRepository<ChatRoom, Long> {}
